+++ js/setup.js
@@ -1,0 +1,1 @@
+export const NUMBER_OF_PHOTOS = 25; // Количество моковых объектов фотографий
